requirements.txt
langchain==0.1.5
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.29.0
faiss-cpu==1.7.4
altair==5.2.0
huggingface-hub
InstructorEmbedding==1.0.1
langchain-community==0.0.17
langchain-core==0.1.18
streamlit-chat==0.1.1
transformers
openai==1.11.1
tiktoken